refactor: 카카오 맵 관련 로직 별도 함수로 분리
